+++ app/src/main/java/exh/md/follows/MangaDexFollowsScreen.kt
@@ -194,8 +194,7 @@ class MangaDexFollowsScreen(private val sourceId: Long) : Screen() {
                     onDismissRequest = onDismissRequest,
                     onEditCategories = { navigator.push(CategoryScreen()) },
                     onConfirm = { include, _ ->
-                        screenModel.changeMangaFavorite(dialog.manga)
-                        screenModel.moveMangaToCategories(dialog.manga, include)
+                        screenModel.confirmCategories(dialog.manga, include, dialog.alreadyFavorited)
                     },
                 )
             }